+++ Resources/Private/JavaScript/asset-variants/package.json
@@ -1,0 +1,9 @@
+{
+  "name": "@media-ui/feature-asset-variants",
+  "version": "1.0.0",
+  "license": "GNU GPLv3",
+  "private": true,
+  "dependencies": {
+    "@media-ui/core": "*"
+  }
+}
