+++ Resources/Private/JavaScript/asset-variants/src/interfaces/AssetVariant.ts
@@ -1,4 +1,4 @@
-import GraphQlEntity from './GraphQLEntity';
+import GraphQlEntity from '@media-ui/core/src/interfaces/GraphQLEntity';
 
 type AssetVariantType = 'AssetVariant';
 type CropInformationType = 'CropInformation';
