+++ Resources/Private/JavaScript/media-module/package.json
@@ -12,12 +12,13 @@
     "@fortawesome/react-fontawesome": "^0.1.13",
     "@friendsofreactjs/react-css-themr": "^4.1.0",
     "@media-ui/core": "*",
+    "@media-ui/feature-asset-preview": "*",
+    "@media-ui/feature-asset-upload": "*",
     "@media-ui/feature-asset-usage": "*",
+    "@media-ui/feature-asset-variants": "*",
     "@media-ui/feature-clipboard": "*",
     "@media-ui/feature-concurrent-editing": "*",
     "@media-ui/feature-similar-assets": "*",
-    "@media-ui/feature-asset-preview": "*",
-    "@media-ui/feature-asset-upload": "*",
     "@neos-project/react-ui-components": "^5.3.4",
     "apollo-upload-client": "^14.1.3",
     "classnames": "^2.2.6",
